feat: Improve duplicate and delete UX and UI

+++ app_feup/lib/view/Pages/class_registration_schedule_editor_view.dart
@@ -190,7 +190,8 @@ class _ClassRegistrationScheduleEditorViewState
   final CourseUnitsForClassRegistration courseUnits;
   final ScheduleOption scheduleOption;
   final AppPlannedScheduleDatabase db = AppPlannedScheduleDatabase();
-  
+  bool optionHasChanges = false;
+
   TextEditingController _renameController;
   PageController _pageController;
   List<PageStorageKey<CourseUnit>> _expandableKeys;
@@ -235,6 +236,7 @@ class _ClassRegistrationScheduleEditorViewState
             Expanded(
               child: TextField(
                 onChanged: (text) {
+                  optionHasChanges = true;
                   scheduleOption.name = text;
                 },
                 decoration: InputDecoration(
@@ -252,27 +254,45 @@ class _ClassRegistrationScheduleEditorViewState
             IconButton(
               color: Theme.of(context).accentColor,
               icon: Icon(Icons.file_copy_outlined),
-              onPressed: scheduleOption.isNew()? null : () {
-                // TODO if changes make a confirmation window
-                Navigator.pop(context);
-                Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                        builder: (context) =>
-                            ClassRegistrationScheduleEditorPageView(
-                                ScheduleOption.copy(scheduleOption)
-                            )
-                    )
-                );
-              },
+              onPressed: (scheduleOption.isNew() || optionHasChanges)
+                  ? null
+                  : () {
+                      Navigator.pop(context);
+                      Navigator.push(
+                          context,
+                          MaterialPageRoute(
+                              builder: (context) =>
+                                  ClassRegistrationScheduleEditorPageView(
+                                      ScheduleOption.copy(scheduleOption))));
+                    },
             ),
             IconButton(
               color: Theme.of(context).accentColor,
               icon: Icon(Icons.delete_outline),
-              onPressed: () async {
-                db.deleteOption(scheduleOption);
-                Navigator.pop(context);
-              },
+              onPressed: scheduleOption.isNew()
+                  ? null
+                  : () => showDialog<String>(
+                        context: context,
+                        builder: (BuildContext context) => AlertDialog(
+                          title: const Text('Apagar horário'),
+                          content: const Text(
+                              'Tem a certeza que pretende apagar o horário?'),
+                          actions: <Widget>[
+                            TextButton(
+                              onPressed: () => Navigator.pop(context),
+                              child: const Text('Cancelar'),
+                            ),
+                            TextButton(
+                              onPressed: () {
+                                db.deleteOption(scheduleOption);
+                                Navigator.pop(context);
+                                Navigator.pop(context);
+                              },
+                              child: const Text('Apagar'),
+                            ),
+                          ],
+                        ),
+                      ),
             ),
           ],
         ),
@@ -310,10 +330,9 @@ class _ClassRegistrationScheduleEditorViewState
   }
 
   Widget buildScheduleDisplay(BuildContext context) {
-    final List<Lecture> lectures = scheduleOption.
-      getLectures(_selectedDay, courseUnits.selected);
-    final List<bool> hasDiscontinuity =
-        Lecture.getDiscontinuities(lectures);
+    final List<Lecture> lectures =
+        scheduleOption.getLectures(_selectedDay, courseUnits.selected);
+    final List<bool> hasDiscontinuity = Lecture.getDiscontinuities(lectures);
     final List<bool> hasCollision = Lecture.getCollisions(lectures);
 
     int daysInWeek;
@@ -379,22 +398,20 @@ class _ClassRegistrationScheduleEditorViewState
                     },
                     labelType: NavigationRailLabelType.none,
                     destinations: [
-                      for (int day = 0;
-                          day < daysInWeek;
-                          day++)
+                      for (int day = 0; day < daysInWeek; day++)
                         NavigationRailDestination(
                           icon: getNavigationRailDestinationIcon(
                             context,
                             day,
-                            scheduleOption
-                                .hasCollisions(day, courseUnits.selected),
+                            scheduleOption.hasCollisions(
+                                day, courseUnits.selected),
                             false,
                           ),
                           selectedIcon: getNavigationRailDestinationIcon(
                             context,
                             day,
-                            scheduleOption
-                                .hasCollisions(day, courseUnits.selected),
+                            scheduleOption.hasCollisions(
+                                day, courseUnits.selected),
                             true,
                           ),
                           label: Placeholder(),
@@ -485,6 +502,7 @@ class _ClassRegistrationScheduleEditorViewState
                   selectedTileColor: Theme.of(context).accentColor,
                   onTap: () {
                     setState(() {
+                      optionHasChanges = true;
                       scheduleOption.classesSelected[courseUnit.abbreviation] =
                           courseUnitClass.name;
                     });
